+++ src/components/InstancesList/InstancesList.test.js
@@ -3,7 +3,7 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import { noop } from 'lodash';
 import userEvent from '@testing-library/user-event';
 
-import { screen, fireEvent } from '@testing-library/react';
+import { screen, fireEvent, waitFor } from '@testing-library/react';
 
 import '../../../test/jest/__mock__';
 
@@ -93,6 +93,7 @@ const renderInstancesList = ({ segment }) => {
             segment={segment}
             searchableIndexes={indexes}
             searchableIndexesES={indexesES}
+            fetchFacets={noop}
           />
         </ModuleHierarchyProvider>
       </StripesContext.Provider>